LPS-55886 Move util logic intro Spring bean
