+++ portal-service/src/com/liferay/portal/kernel/search/SearchResultManagerUtil.java
@@ -15,19 +15,7 @@
 package com.liferay.portal.kernel.search;
 
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.util.PortalUtil;
-import com.liferay.portlet.documentlibrary.model.DLFileEntry;
-import com.liferay.portlet.documentlibrary.model.DLFileEntryConstants;
-import com.liferay.portlet.documentlibrary.service.DLAppLocalServiceUtil;
-import com.liferay.portlet.messageboards.model.MBMessage;
-import com.liferay.portlet.messageboards.service.MBMessageLocalServiceUtil;
-import com.liferay.registry.ServiceReference;
-import com.liferay.registry.collections.ServiceReferenceMapper;
-import com.liferay.registry.collections.ServiceTrackerCollections;
-import com.liferay.registry.collections.ServiceTrackerMap;
+import com.liferay.portal.kernel.security.pacl.permission.PortalRuntimePermission;
 
 import java.util.Locale;
 
@@ -42,7 +30,14 @@ public class SearchResultManagerUtil {
 	public static SearchResult createSearchResult(Document document)
 		throws PortalException {
 
-		return _instance._createSearchResult(document);
+		return getSearchResultManager().createSearchResult(document);
+	}
+
+	public static SearchResultManager getSearchResultManager() {
+		PortalRuntimePermission.checkGetBeanProperty(
+			SearchResultManagerUtil.class);
+
+		return _searchResultManager;
 	}
 
 	public static void updateSearchResult(
@@ -50,176 +45,18 @@ public class SearchResultManagerUtil {
 			PortletRequest portletRequest, PortletResponse portletResponse)
 		throws PortalException {
 
-		_instance._updateSearchResult(
+		getSearchResultManager().updateSearchResult(
 			searchResult, document, locale, portletRequest, portletResponse);
 	}
 
-	private SearchResultManagerUtil() {
-		_serviceTrackerMap.open();
+	public void setSearchResultManager(
+		SearchResultManager searchResultManager) {
+
+		PortalRuntimePermission.checkSetBeanProperty(getClass());
+
+		_searchResultManager = searchResultManager;
 	}
 
-	private SearchResult _createSearchResult(Document document)
-		throws PortalException {
-
-		SearchResultManager searchResultManager = _getSearchResultManager(
-			document);
-
-		return searchResultManager.createSearchResult(document);
-	}
-
-	private SearchResultManager _getSearchResultManager(Document document) {
-		String entryClassName = GetterUtil.getString(
-			document.get(Field.ENTRY_CLASS_NAME));
-
-		SearchResultManager searchResultManager = _serviceTrackerMap.getService(
-			entryClassName);
-
-		if (searchResultManager == null) {
-			if (entryClassName.equals(DLFileEntryConstants.getClassName())) {
-				return new DLFileEntrySearchResultManager();
-			}
-			else if (entryClassName.equals(MBMessage.class.getName())) {
-				return new MBMessageSearchResultManager();
-			}
-			else {
-				return new DefaultSearchResultManagerImpl();
-			}
-		}
-
-		return searchResultManager;
-	}
-
-	private void _updateSearchResult(
-			SearchResult searchResult, Document document, Locale locale,
-			PortletRequest portletRequest, PortletResponse portletResponse)
-		throws PortalException {
-
-		SearchResultManager searchResultManager = _getSearchResultManager(
-			document);
-
-		searchResultManager.updateSearchResult(
-			searchResult, document, locale, portletRequest, portletResponse);
-	}
-
-	private static final SearchResultManagerUtil _instance =
-		new SearchResultManagerUtil();
-
-	private final ServiceTrackerMap<String, SearchResultManager>
-		_serviceTrackerMap = ServiceTrackerCollections.singleValueMap(
-			SearchResultManager.class, "(model.className=*)",
-			new ServiceReferenceMapper<String, SearchResultManager>() {
-
-				@Override
-				public void map(
-					ServiceReference<SearchResultManager> serviceReference,
-					Emitter<String> emitter) {
-
-					Object modelClassName = serviceReference.getProperty(
-						"model.className");
-
-					emitter.emit((String)modelClassName);
-				}
-
-			});
-
-	private static class DefaultSearchResultManagerImpl
-		extends BaseSearchResultManager {
-
-		@Override
-		public SearchResult createSearchResult(Document document) {
-			String entryClassName = GetterUtil.getString(
-				document.get(Field.ENTRY_CLASS_NAME));
-			long entryClassPK = GetterUtil.getLong(
-				document.get(Field.ENTRY_CLASS_PK));
-
-			return new SearchResult(entryClassName, entryClassPK);
-		}
-
-		@Override
-		public void updateSearchResult(
-				SearchResult searchResult, Document document, Locale locale,
-				PortletRequest portletRequest, PortletResponse portletResponse)
-			throws PortalException {
-
-			String entryClassName = GetterUtil.getString(
-				document.get(Field.ENTRY_CLASS_NAME));
-			long entryClassPK = GetterUtil.getLong(
-				document.get(Field.ENTRY_CLASS_PK));
-
-			searchResult.setSummary(
-				getSummary(
-					document, entryClassName, entryClassPK, locale,
-					portletRequest, portletResponse));
-		}
-
-	}
-
-	private static class DLFileEntrySearchResultManager
-		extends BaseSearchResultManager {
-
-		@Override
-		protected void addRelatedModel(
-				SearchResult searchResult, Document document, Locale locale,
-				PortletRequest portletRequest, PortletResponse portletResponse)
-			throws PortalException {
-
-			long entryClassPK = GetterUtil.getLong(
-				document.get(Field.ENTRY_CLASS_PK));
-
-			FileEntry fileEntry = DLAppLocalServiceUtil.getFileEntry(
-				entryClassPK);
-
-			if (fileEntry != null) {
-				Summary summary = getSummary(
-					document, DLFileEntry.class.getName(),
-					fileEntry.getFileEntryId(), locale, portletRequest,
-					portletResponse);
-
-				if (Validator.isNull(summary.getContent())) {
-					summary.setContent(fileEntry.getTitle());
-				}
-
-				searchResult.addFileEntry(fileEntry, summary);
-			}
-			else {
-				long classNameId = GetterUtil.getLong(
-					document.get(Field.CLASS_NAME_ID));
-				long classPK = GetterUtil.getLong(document.get(Field.CLASS_PK));
-
-				String className = PortalUtil.getClassName(classNameId);
-
-				Summary summary = getSummary(
-					document, className, classPK, locale, portletRequest,
-					portletResponse);
-
-				searchResult.setSummary(summary);
-			}
-		}
-
-	}
-
-	private static class MBMessageSearchResultManager
-		extends BaseSearchResultManager {
-
-		@Override
-		protected void addRelatedModel(
-				SearchResult searchResult, Document document, Locale locale,
-				PortletRequest portletRequest, PortletResponse portletResponse)
-			throws PortalException {
-
-			long entryClassPK = GetterUtil.getLong(
-				document.get(Field.ENTRY_CLASS_PK));
-
-			MBMessage mbMessage = MBMessageLocalServiceUtil.getMessage(
-				entryClassPK);
-
-			Summary summary = new Summary(null, mbMessage.getBody());
-
-			summary.setEscape(false);
-
-			searchResult.addComment(mbMessage, summary);
-		}
-
-	}
+	private static SearchResultManager _searchResultManager;
 
 }
